backend/nlp-backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# ML Libraries
transformers==4.35.2
torch==2.1.0
numpy==1.26.2
protobuf==4.25.1

# Security & Rate Limiting
slowapi==0.1.9
python-dotenv==1.0.0
